feat: Add RabbitMqInitializer to create queues
Add RabbitMqInitializer class to initialize RabbitMQ queues on application startup. The RabbitMqInitializer class creates two queues, one with the name specified in the "rabbitmq.queue.name" property and another with the name specified in the "rabbitmq.queue.json.name" property. The queues are created using the AmqpAdmin bean provided in the constructor. The RabbitMqInitializer class is annotated with @Component and has a @PostConstruct method createQueues() which is called after the bean is constructed to create the queues.