feat: Add Messaging RabbitMQ README
Add a README file with an overview of the messaging system based on the RabbitMQ message broker. The system includes a producer and a consumer, where the producer sends messages to a RabbitMQ queue and the consumer listens to the queue and consumes the messages.

Also, provide configuration details including the required environment variables and their values. Additionally, mention the implementation steps and how to run the application using Docker containers. Highlight the possible delay in the RabbitMQ container startup and the need to wait for the connection to be established.

Include a section on testing where an HTTP client can be used to send a request and verify the message in the consumer's log output.